requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
billiard==4.2.1
    # via celery
celery==5.4.0
    # via -r requirements.in
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
django==5.1.1
    # via
    #   django-redis
    #   djangorestframework
    #   drf-yasg
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.4
    # via coreschema
kombu==5.4.2
    # via celery
markupsafe==2.1.5
    # via jinja2
packaging==24.1
    # via
    #   drf-yasg
    #   gunicorn
prompt-toolkit==3.0.48
    # via click-repl
psycopg[binary,c,pool]==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3
    # via psycopg
psycopg-c==3.2.3
    # via psycopg
psycopg-pool==3.2.3
    # via psycopg
python-dateutil==2.9.0.post0
    # via celery
pytz==2024.2
    # via drf-yasg
pyyaml==6.0.2
    # via drf-yasg
redis==5.2.0
    # via django-redis
requests==2.32.3
    # via coreapi
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via
    #   psycopg
    #   psycopg-pool
tzdata==2024.2
    # via
    #   celery
    #   kombu
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==2.2.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
